bug fixed

+++ src/main/java/com/example/user_management_service/model/dto/RecordWorkPlaceStatsDTO.java
@@ -15,6 +15,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class RecordWorkPlaceStatsDTO {
     private Field field;
-    private int doctorsByDB;
-    private int doctorsInFact;
+    private long doctorsByDB;
+    private long doctorsInFact;
 }
